+++ socket/src/decorators/class.ts
@@ -1,21 +1,6 @@
 import { getMeta } from '../meta';
 
 /**
- * Creates server with options
- * @param {number | string | Object} serverOrPort
- * @param {Object} options
- * @returns {(target:Function)=>void}
- * @constructor
- */
-export let Connect = (serverOrPort: any, options?: any): ClassDecorator  => {
-  return (target: Function): void => {
-    let meta: SocketIOMeta = getMeta(target.prototype);
-    meta.serverOrPort = serverOrPort;
-    meta.options = options;
-  }
-};
-
-/**
  * Registers middleware
  * @param {Function} fn
  * @returns {(target:Function)=>void}